design 메인, curation, llm 로딩 화면 정리 및 엔드포인트 세팅
