+++ src/main/resources/templates/restaurant/search.html
@@ -1,33 +1,44 @@
 <!DOCTYPE html>
 <html lang="ko" xmlns:layout="http://www.ultraq.net.nz/thymeleaf/layout"
-      layout:decorate="~{layout}">
+      layout:decorate="~{layout}" xmlns:sec="http://www.w3.org/1999/xhtml">
 <head>
   <meta charset="UTF-8">
-  <title>식당 리스트</title>
-  <link rel="stylesheet" href="/css/restaurantList.css">
+  <title>식당 검색</title>
+  <link rel="stylesheet" href="/css/search.css">
 </head>
 <body>
 <main layout:fragment="content">
 
-  <p class="restaurant-title">🍽️ 식당 검색 결과</p>
-
-  <div th:if="${#lists.isEmpty(results)}" class="no-result">
-    검색 결과가 없습니다.
+  <!-- 로그인한 사용자 정보 확인용 -->
+  <div sec:authorize="isAuthenticated()">
+    <span class="text-sm font-semibold">
+      <span th:text="${#authentication.principal.user.nickname}">닉네임</span> 님 환영합니다!
+    </span>
   </div>
-  <ul th:if="${not #lists.isEmpty(results)}" class="restaurant-list">
-  <ul class="restaurant-list">
-    <li th:each="res : ${results}" class="restaurant-card">
-      <a th:href="@{/search/detail/{id}(id=${res.id})}">
-        <div>
-          <span class="restaurant-name" th:text="${res.name}">식당명</span>
-          <span class="restaurant-branch" th:text="${res.branch}">지점명</span>
-          <span class="restaurant-category" th:text="${res.category}">한식</span>
-        </div>
-        <p class="restaurant-time" th:text="${res.businessTime}">영업시간</p>
-        <p class="restaurant-address" th:text="${res.roadAddress}">주소</p>
-      </a>
-    </li>
-  </ul>
+
+  <div class="search-box">
+  <p class="main-title">🍽️ 찾으시는 식당이 있으신가요?</p>
+
+  <form method="get" action="/search/result" class="search" onsubmit="return validateForm()">
+    <input type="text" name="restaurantKeyword" id="restaurantKeyword" placeholder="식당명을 입력하세요">
+    <button type="submit">검색</button>
+  </form>
+  </div>
+
+
+
+  <script>
+    function validateForm(){
+      const keyword = document.getElementById('restaurantKeyword').value.trim();
+      if(!keyword){
+        alert("식당명을 입력해 주세요.");
+        return false;
+      }
+      return true;
+    }
+
+  </script>
+
 </main>
 </body>
 </html>
